Update strategy_registry logic to avoid duplication code to register strategies
